+++ settings.js
@@ -8,9 +8,9 @@ const SETTINGS = {};
 SETTINGS['watcha.com'] = {
   includingPaths: ['/browse/video', '/tag', '/explore', '/watched', '/library', '/watchings', '/ratings', '/arrivals', '/staffmades', '/contents', '/people', '/content_decks'],
   rootSelector: 'main',
-  selector: `section:not([class$="BrowseSection"]) ul>li>div[class*="-Cell"]:not([${FU}])>a[class][href^="/contents/"]>div:not(:has(>figure)), `  //list item
+  selector: `section:not([class$="BrowseSection"])>div>ul>li>div[class*="-Cell"]:not([${FU}])>a[class][href^="/contents/"]>div:not(:has(>figure)), `  //list item
   + `section>ul>li:not([${FU}])>a[href^="/contents/"]>div, `  //search
-  + `section ul>li>article:not([${FU}])>a[href^="/contents/"]>div`,   //list item on single-page
+  + `section ul>li>article:not([class]):not([${FU}])>a[href^="/contents/"]>div`,   //list item on single-page or libray
 
   //numberToBaseEl: 2,  //when edit, this number + 1 is used (old)